requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1

langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

chromadb>=0.5.0
tiktoken>=0.8.0
langchain-text-splitters>=0.3.0

matplotlib>=3.8.0
pandas>=2.2.0
